collected all relations in one file. updated imports
