--- src/db/relations/commentRelations.ts
@@ -1,11 +1,0 @@
-import { relations } from "drizzle-orm";
-import { comments } from "../tables/comments";
-
-export const commentRelations = relations(comments, ({ one, many }) => ({
-  parent: one(comments, {
-    fields: [comments.parentId],
-    references: [comments.id],
-    relationName: 'comment_parent',
-  }),
-  replies: many(comments, { relationName: 'comment_parent' }),
-}));
